python/requirements.txt
Flask==2.2.5
web3==6.5.0
joblib==1.3.2
numpy==1.24.2
scikit-learn==1.2.2
python-dotenv==1.0.0
